requirements.txt
appium-python-client
pytest
allure-pytest